refactor: Extract user functionality from AuthDataSource into UserDataSource.
