+++ app/src/main/java/no/usn/mob3000/domain/source/IAuthDataSource.kt
@@ -15,9 +15,4 @@ import no.usn.mob3000.data.model.social.ProfileDto
  */
 interface IAuthDataSource {
     suspend fun getCurrentSession(): UserSession?
-    suspend fun getCurrentUser(): UserInfo
-    suspend fun getUserGameStats(): GameDataDto
-    suspend fun getUserProfile(userId: String): ProfileDto?
-    suspend fun getUserFriends(userId: String): List<FriendsDto>
-    suspend fun getUserFriendRequests(): List<FriendRequestsDto>
 }
