+++ app/src/main/java/no/usn/mob3000/domain/usecase/auth/LogoutUseCase.kt
@@ -2,6 +2,7 @@ package no.usn.mob3000.domain.usecase.auth
 
 import no.usn.mob3000.data.repository.auth.AuthRepository
 import no.usn.mob3000.data.source.remote.auth.AuthDataSource
+import no.usn.mob3000.data.source.remote.auth.UserDataSource
 
 /**
  * Android Use Case for handling logout operations.
@@ -11,7 +12,7 @@ import no.usn.mob3000.data.source.remote.auth.AuthDataSource
  * @created 2024-10-25
  */
 class LogoutUseCase(
-    private val authRepository: AuthRepository = AuthRepository(AuthDataSource())
+    private val authRepository: AuthRepository = AuthRepository(AuthDataSource(), UserDataSource())
 ) {
     suspend operator fun invoke(): Result<Unit> = try {
         authRepository.logout()
