+++ app/src/main/java/no/usn/mob3000/domain/usecase/auth/RegisterUseCase.kt
@@ -2,6 +2,7 @@ package no.usn.mob3000.domain.usecase.auth
 
 import no.usn.mob3000.data.repository.auth.AuthRepository
 import no.usn.mob3000.data.source.remote.auth.AuthDataSource
+import no.usn.mob3000.data.source.remote.auth.UserDataSource
 
 /**
  * Android Use Case for handling registration operations.
@@ -15,7 +16,7 @@ import no.usn.mob3000.data.source.remote.auth.AuthDataSource
  * @created 2024-10-30
  */
 class RegisterUseCase(
-    private val authRepository: AuthRepository = AuthRepository(AuthDataSource())
+    private val authRepository: AuthRepository = AuthRepository(AuthDataSource(), UserDataSource())
 ) {
     suspend operator fun invoke(email: String, password: String): Result<Unit> {
         if (!email.isValidEmail() || password.length < 8) {
